proxy: Fix a bug that Listener doesn't release the port when closed.
A simpler fix is just to add the goroutine in Serve():
```
go func() {
  <-l.stopChan
  l.connWG.Wait()
  listen.Close()
}()
```

This is almost enough, but there's a subtle race between the call to
`l.Close()` and the port actually being released: they both wait for
all of the connections to finish, but then the return from `l.Close()`
and the call to `listen.Close()` in `Serve` are racy.

While this is unlikely to matter in the vast majority of cases, the
addition of `stoppedChan` fixes this completely. Now a call to
`l.Close()` will not return on the listener has been stopped and the
bound port has been released.

This ensures that the added test is not flaky.